Added reporters to docstring of simulation class (#52)
Problem:
The reporters argument was not mentioned in the docstring of simulation.Simulation class. 
This meant that it is currently not visible in the argument is invisible in the readthedocs.

Proposed solution: 
Added
'''
reporters: list, optional 
            List of reporters to use in the simulation.
'''

Feel free to suggest alterations to the docstring entry if you do not find it to describe what you want it to.